backend/requirements.txt
flask
flask-cors
requests
werkzeug
rapidfuzz
numpy
faiss-cpu
sentence-transformers
google-generativeai
